requirement.txt
python-Levenshtein==0.12.0
python=3.6.8
pillow
opencv-python
torch==1.0.1
torchvision
tensorflow==1.7.0
lmdb
easydict
pyyaml